ptsd-model-api/requirements.txt
flask
torch
pillow
torchvision